requirements.txt
flask==3.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow>=11.0.0,<12
transformers>=4.35.0,<5.0.0
torch>=2.6.0,<2.9.0
sentence-transformers>=2.2.2,<3.0.0
werkzeug==3.0.1
python-docx==1.1.0
pandas>=2.1.3,<3.0.0
requests>=2.31.0,<3.0.0
